Created variables for EnemyAIBase
Created enemy variables for the enemy’s current health, max health, enemy speed, enemies’ detection mesh, a transform object for the player, and the bool that toggles when the player is in range of the enemy. Instantiated and assigned enemy current health to max health and the transform player object var is assigned to find the player string tag. All methods in enemy AI base will be virtual as this will be the parent for all enemy AI scripts; their methods will have to be set to override.